requirements.txt
# OpenAI SDK
openai>=1.50.0

# Environment Management
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Excel Support
openpyxl>=3.1.0

# Web UI Framework
gradio>=4.0.0

# Vector Search (for RAG)
faiss-cpu>=1.7.4

# Optional: For notebook support
# jupyter>=1.0.0
# ipykernel>=6.25.0

# Optional: For testing
# pytest>=7.4.0
